decorator

+++ Model/Bevanda.java
@@ -1,5 +1,5 @@
 package Model;
 interface Bevanda {
     public String getDescrizione();
-    public int getCosto();
+    public double getCosto();
 }
